Add files via upload
This analysis is about the current twitter climate for the European Super League, a new format soccer league for the best and most popular clubs in soccer. Looking at what the Super League can focus on to build a better campaign for an eventual relaunch.

The PR battle that followed the announcement of the European super league, will be a text book case in a few years. The mightiest most popular brands in sports were taken down by a social media and media campaign within two days. The whole battle was over before we managed to learn what the majority actually wanted. Now almost half a year later we can look at the sentiments of people now that the initial storm has calmed.